feat(layout): 💄 reorganize `Header` and `Footer` components
