+++ package.json
@@ -38,6 +38,7 @@
   },
   "dependencies": {
     "@inlang/paraglide-sveltekit": "^0.16.0",
+    "@lucide/svelte": "^0.479.0",
     "@tailwindcss/typography": "^0.5.16"
   }
 }
