+++ src/routes/+page.svelte
@@ -1,177 +1,39 @@
 <script lang="ts">
-  import { goto } from '$app/navigation';
-  import { page } from '$app/state';
-  import { i18n } from '$lib/i18n';
   import * as m from '$lib/paraglide/messages.js';
-  import type { AvailableLanguageTag } from '$lib/paraglide/runtime';
+  </script>
 
-  let isMobileMenuOpen = $state(false);
-
-  const navigation = [
-    { name: 'Qui som', href: '#qui-som' },
-    { name: 'Trajectòria', href: '#trajectòria' },
-    { name: 'Concerts', href: '#concerts' },
-    { name: 'Contacte', href: '#contacte' }
-  ];
-  const languages: { name: string; code: AvailableLanguageTag }[] = [
-    { name: 'Català', code: 'ca' },
-    { name: 'Español', code: 'es' },
-    { name: 'English', code: 'en' }
-  ];
-
-  function switchToLanguage(newLanguage: AvailableLanguageTag) {
-    const canonicalPath = i18n.route(page.url.pathname);
-    const localisedPath = i18n.resolveRoute(canonicalPath, newLanguage);
-    goto(localisedPath);
-  }
-</script>
-
-<div class="bg-white">
-  <header class="absolute inset-x-0 top-0 z-50">
-    <nav class="flex items-center justify-between p-6 lg:px-8" aria-label="Global">
-      <div class="flex lg:flex-1">
-        <a href="#" class="-m-1.5 p-1.5">
-          <span class="sr-only">Coral Jove del Conservatori de Sant Cugat</span>
-          <img
-            class="h-10 w-auto"
-            src="/logo.png"
-            alt="Logo Coral Jove del Conservatori de Sant Cugat"
-          />
-        </a>
-      </div>
-      <div class="flex lg:hidden">
-        <button
-          type="button"
-          class="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
-          onclick={() => (isMobileMenuOpen = true)}
+<div class="relative isolate px-6 pt-14 lg:px-8">
+  <div
+    class="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
+    aria-hidden="true"
+  >
+    <div
+      class="relative left-[calc(50%-11rem)] aspect-1155/678 w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-linear-to-tr from-[var(--primary)] to-[#10a65a] opacity-30 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
+      style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)"
+    ></div>
+  </div>
+  <div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56">
+    <div class="text-center">
+      <h1 class="text-5xl font-semibold tracking-tight text-balance text-gray-900 sm:text-7xl">
+        Coral Jove del Conservatori de Sant Cugat
+      </h1>
+      <p class="mt-8 text-lg font-medium text-pretty text-gray-500 sm:text-xl/8">
+        Som un grup coral de joves d’entre 20 i 25 anys dirigits per Raimon Romaní
+      </p>
+      <div class="mt-10 flex items-center justify-center gap-x-6">
+        <a href="#qui-som" class="text-sm/6 font-semibold text-gray-900"
+          >{m.aboutUs()} <span aria-hidden="true">→</span></a
         >
-          <span class="sr-only">Open main menu</span>
-          <svg
-            class="size-6"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke-width="1.5"
-            stroke="currentColor"
-            aria-hidden="true"
-            data-slot="icon"
-          >
-            <path
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5"
-            />
-          </svg>
-        </button>
       </div>
-      <div class="hidden lg:flex lg:gap-x-12">
-        {#each navigation as { name, href }}
-          <a {href} class="text-sm/6 font-semibold text-gray-900">{name}</a>
-        {/each}
-      </div>
-      <div class="hidden gap-2 text-sm lg:flex lg:flex-1 lg:justify-end">
-        {#each languages as { name, code }}
-          <button onclick={() => switchToLanguage(code)}>{name}</button>
-        {/each}
-      </div>
-    </nav>
-    <!-- Mobile menu, show/hide based on menu open state. -->
-    <dialog
-      class="lg:hidden"
-      aria-modal="true"
-      open={isMobileMenuOpen}
-      onclose={() => (isMobileMenuOpen = false)}
-    >
-      <!-- Background backdrop, show/hide based on slide-over state. -->
-      <div class="fixed inset-0 z-50"></div>
-      <div
-        class="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10"
-      >
-        <div class="flex items-center justify-between">
-          <a href="#" class="-m-1.5 p-1.5">
-            <span class="sr-only">Coral Jove del Conservatori de Sant Cugat</span>
-            <img
-              class="h-10 w-auto"
-              src="/logo.png"
-              alt="Logo Coral Jove del Conservatori de Sant Cugat"
-            />
-          </a>
-          <button
-            type="button"
-            class="-m-2.5 rounded-md p-2.5 text-gray-700"
-            onclick={() => (isMobileMenuOpen = false)}
-          >
-            <span class="sr-only">Tanca el menú</span>
-            <svg
-              class="size-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke-width="1.5"
-              stroke="currentColor"
-              aria-hidden="true"
-              data-slot="icon"
-            >
-              <path stroke-linecap="round" stroke-linejoin="round" d="M6 18 18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-        <div class="mt-6 flow-root">
-          <div class="-my-6 divide-y divide-gray-500/10">
-            <div class="space-y-2 py-6">
-              {#each navigation as { name, href }}
-                <a
-                  {href}
-                  class="-mx-3 block rounded-lg px-3 py-2 text-base/7 font-semibold text-gray-900 hover:bg-gray-50"
-                  >{name}</a
-                >
-              {/each}
-            </div>
-            <div class="py-6">
-              {#each languages as { name, code }}
-                <button
-                  class="-mx-3 block rounded-lg px-3 py-2.5 text-base/7 font-semibold text-gray-900 hover:bg-gray-50"
-                  onclick={() => switchToLanguage(code)}>{name}</button
-                >
-              {/each}
-            </div>
-          </div>
-        </div>
-      </div>
-    </dialog>
-  </header>
-
-  <div class="relative isolate px-6 pt-14 lg:px-8">
+    </div>
+  </div>
+  <div
+    class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
+    aria-hidden="true"
+  >
     <div
-      class="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
-      aria-hidden="true"
-    >
-      <div
-        class="relative left-[calc(50%-11rem)] aspect-1155/678 w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-linear-to-tr from-[var(--primary)] to-[#10a65a] opacity-30 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
-        style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)"
-      ></div>
-    </div>
-    <div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56">
-      <div class="text-center">
-        <h1 class="text-balance text-5xl font-semibold tracking-tight text-gray-900 sm:text-7xl">
-          Coral Jove del Conservatori de Sant Cugat
-        </h1>
-        <p class="mt-8 text-pretty text-lg font-medium text-gray-500 sm:text-xl/8">
-          Som un grup coral de joves d’entre 20 i 25 anys dirigits per Raimon Romaní
-        </p>
-        <div class="mt-10 flex items-center justify-center gap-x-6">
-          <a href="#qui-som" class="text-sm/6 font-semibold text-gray-900"
-            >{m.aboutUs()} <span aria-hidden="true">→</span></a
-          >
-        </div>
-      </div>
-    </div>
-    <div
-      class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
-      aria-hidden="true"
-    >
-      <div
-        class="relative left-[calc(50%+3rem)] aspect-1155/678 w-[36.125rem] -translate-x-1/2 bg-linear-to-tr from-[var(--primary)] to-[#10a65a] opacity-30 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
-        style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)"
-      ></div>
-    </div>
+      class="relative left-[calc(50%+3rem)] aspect-1155/678 w-[36.125rem] -translate-x-1/2 bg-linear-to-tr from-[var(--primary)] to-[#10a65a] opacity-30 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
+      style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)"
+    ></div>
   </div>
 </div>
